+++ packages/gamut/src/Form/SelectDropdown/index.tsx
@@ -1,2 +1,6 @@
+export type {
+  OptionStrict,
+  ExtendedOption,
+  SelectDropdownGroup,
+} from './types/options';
 export * from './SelectDropdown';
-export * from './types';